add: member and edit member pages

+++ src/server/api/routers/membership.ts
@@ -1,10 +1,10 @@
-import type { Account } from "~/server/db/schema";
+import type { Account, Membership } from "~/server/db/schema";
 import { memberships } from "~/server/db/schema";
 import { z } from "zod";
-import { adminProcedure, createTRPCRouter, loggedInProcedure, userProcedure } from "~/server/api/trpc";
+import { accessedProcedure, adminProcedure, createTRPCRouter, loggedInProcedure, userProcedure } from "~/server/api/trpc";
 import { TRPCError } from "@trpc/server";
 import { and, eq } from "drizzle-orm";
-import { Access } from "~/utils/types";
+import { Access, Member } from "~/utils/types";
 import { nanoid } from "nanoid";
 
 type GetAccountType = {
@@ -13,6 +13,38 @@ type GetAccountType = {
 };
 
 export const membershipRouter = createTRPCRouter({
+  get: accessedProcedure
+    .input(z.object({
+      membershipId: z.string(),
+    }))
+    .query(async ({ input, ctx }): Promise<Member | null> => {
+      const { membershipId } = input;
+
+      let membership: Member | undefined;
+      try {
+        membership = await ctx.db.query.memberships.findFirst({
+          where: eq(memberships.id, membershipId),
+          with: {
+            user: true,
+          },
+        });
+      } catch (e) {
+        throw new TRPCError({
+          code: "INTERNAL_SERVER_ERROR",
+          message: "Membership retrieval failed",
+        })
+      }
+
+      if (!membership) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Membership not found",
+        })
+      }
+
+      return membership;
+    }),
+
   getAccounts: loggedInProcedure
     .query(async ({ ctx }): Promise<GetAccountType[] | null> => {
       const { id: userId } = ctx.self.user;
@@ -38,16 +70,45 @@ export const membershipRouter = createTRPCRouter({
 
   delete: adminProcedure
     .input(z.object({
-      userId: z.string(),
+      membershipId: z.string(),
     }))
     .mutation(async ({ input, ctx }): Promise<true> => {
-      const { userId } = input;
-      const { accountId } = ctx;
+      const { membershipId } = input;
 
+      let membership: Membership | undefined;
+      try {
+        membership = await ctx.db.query.memberships.findFirst({
+          with: {
+            user: true,
+          },
+          where: and(
+            eq(memberships.id, membershipId),
+          ),
+        });
+
+      } catch (e) {
+        throw new TRPCError({
+          code: "INTERNAL_SERVER_ERROR",
+          message: "Membership retrieval failed",
+        })
+      }
+
+      if (!membership) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Membership not found",
+        })
+      }
+
+      if (membership.userId === ctx.self.user.id) {
+        throw new TRPCError({
+          code: "CONFLICT",
+          message: "Cannot delete self",
+        })
+      }
       try {
         await ctx.db.delete(memberships).where(and(
-          eq(memberships.userId, userId),
-          eq(memberships.accountId, accountId),
+          eq(memberships.id, membershipId),
         ));
         return true;
       } catch (e) {
@@ -73,7 +134,7 @@ export const membershipRouter = createTRPCRouter({
           message: "No user found",
         })
       }
-    
+
       if (existingAccess === "admin" || existingAccess === "viewer") {
         throw new TRPCError({
           code: "CONFLICT",
@@ -99,6 +160,59 @@ export const membershipRouter = createTRPCRouter({
         throw new TRPCError({
           code: "INTERNAL_SERVER_ERROR",
           message: "Membership creation failed",
+        })
+      }
+    }),
+
+  update: adminProcedure
+    .input(z.object({
+      membershipId: z.string(),
+      access: z.enum(["admin", "viewer"]),
+    }))
+    .mutation(async ({ input, ctx }): Promise<true> => {
+      const { membershipId, access } = input;
+
+      let membership: Membership | undefined;
+      try {
+        membership = await ctx.db.query.memberships.findFirst({
+          where: and(
+            eq(memberships.id, membershipId),
+          ),
+        });
+
+      } catch (e) {
+        throw new TRPCError({
+          code: "INTERNAL_SERVER_ERROR",
+          message: "Member retrieval failed",
+        })
+      }
+
+      if (!membership) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Member not found",
+        })
+      }
+
+      if (membership.userId === ctx.self.user.id) {
+        throw new TRPCError({
+          code: "CONFLICT",
+          message: "Cannot update self",
+        })
+      }
+
+      try {
+        await ctx.db.update(memberships).set({
+          access,
+        }).where(and(
+          eq(memberships.id, membershipId),
+        ));
+
+        return true;
+      } catch (e) {
+        throw new TRPCError({
+          code: "INTERNAL_SERVER_ERROR",
+          message: "Membership update failed",
         })
       }
     }),
